DAOS-11876 chk: attach pools tree to chk_instance on leader
Originally, the pools tree on check leader was just used to drive
pools check, but not trace each pool's process. But if we want to
support start DAOS check for multiple pools with different phases,
we need to trace each pool's process on check leader. Then we can
share related logic with DAOS check engine on which related pools
tree is attached to the chk_instance. Then we change check leader
as the same to unify and share related logic.

Signed-off-by: Fan Yong <fan.yong@intel.com>